backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
dbus-python==1.3.2
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
dnf==4.21.1
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
libcomps==0.1.20
libdnf==0.73.3
MarkupSafe==3.0.2
nftables==0.1
numpy==2.1.3
opencv-python==4.10.0.84
PyGObject==3.48.2
python-dateutil==2.8.2
python-dotenv==1.0.1
rpm==4.19.1.1
six==1.16.0
sqlparse==0.5.0
systemd-python==235
Werkzeug==3.1.0
